+++ lib/interactor/contracts/breach.rb
@@ -35,6 +35,26 @@ module Interactor
       # @api public
       # @return [Symbol] the property violated in the contract
       attr_reader :property
+
+      # Allows the Breach to be splatted out as arguments to a block
+      #
+      # @api private
+      # @return [Array<Symbol, Array<String>>]
+      def to_ary
+        [property, messages]
+      end
+
+      # Converts the Breach to an equivalent Hash
+      #
+      # @example
+      #   breach = Interactor::Contracts::Breach.new(:name, ["name is missing"])
+      #   breach.to_h  #=> {:name => ["name is missing"]}
+      #
+      # @api public
+      # @return [Hash]
+      def to_h
+        {property => messages}
+      end
     end
   end
 end