+++ lib/interactor/contracts/outcome.rb
@@ -1,5 +1,6 @@
 require "forwardable"
 require "interactor/contracts/breach"
+require "interactor/contracts/breach_set"
 
 module Interactor
   module Contracts
@@ -45,9 +46,9 @@ module Interactor
       # @api semipublic
       # @return [Array<Breach>] the breaches of the Terms' constraints
       def breaches
-        result.messages.map do |property, messages|
+        BreachSet.new(result.messages(:full => true).map do |property, messages|
           Breach.new(property, messages)
-        end
+        end)
       end
 
       # Checks whether the the Terms enforcement was a failure